GeoData

+++ app/src/main/java/com/Ton_in/NoodleMap/GeoData.java
@@ -2,13 +2,13 @@ package com.Ton_in.NoodleMap;
 
 public class GeoData
 {
-	public float lattitude;
-	public float longitude;
+	public double lattitude;
+	public double longitude;
 	public String text;
 	
 	public GeoData()
 	{
-		lattitude = 0.0f;
-		longitude = 0.0f;
+		lattitude = 0.0;
+		longitude = 0.0;
 	}
 }
